Clean up a bunch of code and eslint warnings (#765)
