+++ common/utils/headways.ts
@@ -32,14 +32,14 @@ export const longestHeadway = (headways: SingleDayDataPoint[] | AggregateDataPoi
   }
 };
 
-export const longestAggregateHeadway = (headways: AggregateDataPoint[]) => {
+const longestAggregateHeadway = (headways: AggregateDataPoint[]) => {
   return headways.reduce(
     (current, datapoint) => (datapoint.min < current.min ? datapoint : current),
     headways[0]
   );
 };
 
-export const longestSingleHeadway = (headways: SingleDayDataPoint[]) => {
+const longestSingleHeadway = (headways: SingleDayDataPoint[]) => {
   return headways.reduce((current, datapoint) => {
     if (datapoint.headway_time_sec && current.headway_time_sec)
       return datapoint.headway_time_sec > current.headway_time_sec ? datapoint : current;
@@ -47,7 +47,7 @@ export const longestSingleHeadway = (headways: SingleDayDataPoint[]) => {
   }, headways[0]);
 };
 
-export const shortestSingleHeadway = (headways: SingleDayDataPoint[]) => {
+const shortestSingleHeadway = (headways: SingleDayDataPoint[]) => {
   return headways.reduce((current, datapoint) => {
     if (datapoint.headway_time_sec && current.headway_time_sec)
       return datapoint.headway_time_sec < current.headway_time_sec ? datapoint : current;
